+++ src/components/NavBar/NavBarWallet.test.tsx
